requirements.txt
Django==4.0.3
djangorestframework==3.12.2
psycopg2-binary==2.9.3
celery==5.2.6
django-environ==0.8.1
djangorestframework-simplejwt==4.7.1